obrender/render: make alpha-clear mask unsigned
Build the alpha-clearing mask as a guint32 before applying bitwise NOT,
avoiding implementation-defined behavior from operating on signed ints.

+++ obrender/render.c
@@ -558,7 +558,7 @@ gboolean RrPixmapToRGBA(const RrInstance *inst,
         for (i = 0, y = 0; y < ph; ++y) {
             for (x = 0; x < pw; ++x, ++i) {
                 if (!((((unsigned)xm->data[di + x / 8]) >> (x % 8)) & 0x1))
-                    (*data)[i] &= ~(0xff << RrDefaultAlphaOffset);
+                    (*data)[i] &= ~((guint32)0xff << RrDefaultAlphaOffset);
             }
             di += xm->bytes_per_line;
         }
